+++ R/exportFromSCE.R
@@ -40,7 +40,7 @@
 #' No data returned.
 #'
 #' @examples
-#' pbmc <- readRDS(system.file("extdata/v1.4/pbmc_SCE.rds",
+#' pbmc <- readRDS(system.file("extdata/examples/pbmc_SCE.rds",
 #'   package = "CerebroNexus"))
 #' exportFromSCE(
 #'   object = pbmc,