+++ R/getEnrichedPathways.R
@@ -27,7 +27,7 @@
 #'
 #' @examples
 #' if (interactive()) {
-#'   pbmc <- readRDS(system.file("extdata/v1.4/pbmc_seurat.rds",
+#'   pbmc <- readRDS(system.file("extdata/examples/pbmc_seurat.rds",
 #'     package = "CerebroNexus"))
 #'   pbmc <- getEnrichedPathways(
 #'     object = pbmc,